Move queries and mutations to frontend/src/graphql.
